refactor mediastreaming package converters
