+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/models/MediaStreamingPackageParser.java
@@ -3,8 +3,8 @@
 
 package com.azure.communication.callautomation.models;
 
-import com.azure.communication.callautomation.implementation.models.MediaStreamingAudioDataInternal;
-import com.azure.communication.callautomation.implementation.models.MediaStreamingMetadataInternal;
+import com.azure.communication.callautomation.implementation.converters.MediaStreamingAudioDataConverter;
+import com.azure.communication.callautomation.implementation.converters.MediaStreamingMetadataConverter;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.logging.ClientLogger;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -54,11 +54,11 @@ public class MediaStreamingPackageParser {
             mapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
             JsonNode jsonData = mapper.readTree(stringJson);
             if (stringJson.contains("AudioData")) {
-                MediaStreamingAudioDataInternal audioInternal = mapper.convertValue(jsonData.get("audioData"), MediaStreamingAudioDataInternal.class);
+                MediaStreamingAudioDataConverter audioInternal = mapper.convertValue(jsonData.get("audioData"), MediaStreamingAudioDataConverter.class);
                 return new MediaStreamingAudioData(audioInternal.getData(), audioInternal.getTimestamp(), audioInternal.getParticipantRawID(), audioInternal.isSilent());
             }
             if (stringJson.contains("AudioMetadata")) {
-                MediaStreamingMetadataInternal metadataInternal = mapper.convertValue(jsonData.get("audioMetadata"), MediaStreamingMetadataInternal.class);
+                MediaStreamingMetadataConverter metadataInternal = mapper.convertValue(jsonData.get("audioMetadata"), MediaStreamingMetadataConverter.class);
                 return new MediaStreamingMetadata(metadataInternal.getMediaSubscriptionId(), metadataInternal.getEncoding(), metadataInternal.getSampleRate(), metadataInternal.getChannels(), metadataInternal.getLength());
             }
             return null;
